+++ fiddles/react/fiddle-0017-EnterpriseAgGrid/src/app/App.js
@@ -1,0 +1,10 @@
+import './App.css';
+import RichGrid from './components';
+
+function App() {
+  return (
+      <RichGrid></RichGrid>
+  );
+}
+
+export default App;