+++ fiddles/react/fiddle-0017-EnterpriseAgGrid/src/app/components/index.js
@@ -1,0 +1,4 @@
+
+import RichGrid from './rich-grid/rich-grid.jsx';
+
+export default RichGrid;